pyproject.toml
[tool.poetry]
name = "tt-flashbots-worktest"
version = "0.1.0"
description = "Flashbots worktest project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
pandas = ">=2.2.0"
plotly = "^6.4.0"
httpx = "^0.28.1"
pydantic = "^2.12.4"
sqlalchemy = "^2.0.44"
psycopg = { version = "^3.2.0", extras = ["binary"] }
python-dotenv = "^1.0.0"
greenlet = "^3.2.4"
colorlog = "^6.10.1"
tqdm = "^4.67.1"
rich = "^14.2.0"
pyarrow = "^22.0.0"
websockets = "^15.0"
types-defusedxml = "^0.7.0.20250822"
grafanalib = "^0.7.1"
matplotlib = "^3.10.7"
kaleido = "^1.2.0"
beautifulsoup4 = "^4.14.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = ">=4.1.0"
pytest-timeout = "^2.3.0"
pytest-asyncio = "^1.3.0"
ruff = ">=0.4.0"
pyright = ">=1.1.0"
pandas-stubs = "^2.3.2.250926"
pytest-postgresql = "^7.0.2"
pytest-docker = "^3.2.5"
pytest-httpx = "^0.35.0"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--timeout=10",
    "-m",
    "not integration",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
omit = [
    "src/live.py",
    "src/**/backfill.py",
    "src/**/backfill_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pyright]
# Configuration moved to pyrightconfig.json
